+++ apps/api/src/hyp3_api/handlers.py
@@ -55,7 +55,8 @@ def post_jobs(body, user):
 
     quota = get_user(user)['quota']
     if quota['remaining'] - len(body['jobs']) < 0:
-        message = f'Your monthly quota is {quota["limit"]} jobs. You have {quota["remaining"]} jobs remaining.'
+        max_jobs = quota['max_jobs_per_month']
+        message = f'Your monthly quota is {max_jobs} jobs. You have {quota["remaining"]} jobs remaining.'
         return problem(400, 'Bad Request', message)
 
     try:
@@ -66,7 +67,7 @@ def post_jobs(body, user):
         return problem(400, 'Bad Request', str(e))
 
     request_time = format_time(datetime.now(timezone.utc))
-    table = DYNAMODB_RESOURCE.Table(environ['TABLE_NAME'])
+    table = DYNAMODB_RESOURCE.Table(environ['JOBS_TABLE_NAME'])
 
     for job in body['jobs']:
         job['job_id'] = str(uuid4())
@@ -81,7 +82,7 @@ def post_jobs(body, user):
 
 
 def get_jobs(user, start=None, end=None, status_code=None, name=None):
-    table = DYNAMODB_RESOURCE.Table(environ['TABLE_NAME'])
+    table = DYNAMODB_RESOURCE.Table(environ['JOBS_TABLE_NAME'])
 
     key_expression = Key('user_id').eq(user)
     if start is not None or end is not None:
@@ -102,7 +103,7 @@ def get_jobs(user, start=None, end=None, status_code=None, name=None):
 
 
 def get_job_by_id(job_id):
-    table = DYNAMODB_RESOURCE.Table(environ['TABLE_NAME'])
+    table = DYNAMODB_RESOURCE.Table(environ['JOBS_TABLE_NAME'])
     response = table.get_item(Key={'job_id': job_id})
     if 'Item' not in response:
         return problem(404, 'Not Found', f'job_id does not exist: {job_id}')
@@ -110,7 +111,7 @@ def get_job_by_id(job_id):
 
 
 def get_names_for_user(user):
-    table = DYNAMODB_RESOURCE.Table(environ['TABLE_NAME'])
+    table = DYNAMODB_RESOURCE.Table(environ['JOBS_TABLE_NAME'])
     key_expression = Key('user_id').eq(user)
     response = table.query(
         IndexName='user_id',
@@ -120,12 +121,24 @@ def get_names_for_user(user):
     return sorted(list(names))
 
 
+def get_max_jobs_per_month(user):
+    table = DYNAMODB_RESOURCE.Table(environ['USERS_TABLE_NAME'])
+    response = table.get_item(Key={'user_id': user})
+    if 'Item' in response:
+        max_jobs_per_month = response['Item']['max_jobs_per_month']
+    else:
+        max_jobs_per_month = int(environ['MONTHLY_JOB_QUOTA_PER_USER'])
+    return max_jobs_per_month
+
+
 def get_user(user):
+    max_jobs = get_max_jobs_per_month(user)
+
     return {
         'user_id': user,
         'quota': {
-            'limit': int(environ['MONTHLY_JOB_QUOTA_PER_USER']),
-            'remaining': get_remaining_jobs_for_user(user),
+            'max_jobs_per_month': max_jobs,
+            'remaining': get_remaining_jobs_for_user(user, max_jobs),
         },
         'job_names': get_names_for_user(user)
     }